renamed spotify auth stuff accordingly
