+++ backend/src/auth/auth.module.ts
@@ -11,9 +11,9 @@ import { PrismaModule } from "../../prisma/prisma.module";
 import { UsersService } from "src/modules/users/users.service";
 import { DbUtilsService } from "src/modules/db-utils/db-utils.service";
 import { DtoGenService } from "src/modules/dto-gen/dto-gen.service";
-import { SpotifyAuthService } from "./spotify/spotify.service";
-import { SpotifyAuthController } from "./spotify/spotify.controller";
-import { SpotifyAuthModule } from "./spotify/spotify.module";
+import { SpotifyAuthService } from "./spotify/spotifyauth.service";
+import { SpotifyAuthController } from "./spotify/spotifyauth.controller";
+import { SpotifyAuthModule } from "./spotify/spotifyauth.module";
 import { HttpModule } from "@nestjs/axios";
 
 @Module({
