+++ backend/src/auth/spotify/spotifyauth.module.ts
@@ -1,6 +1,6 @@
 import { HttpModule } from "@nestjs/axios";
 import { Module } from "@nestjs/common";
-import { SpotifyAuthService } from "./spotify.service";
+import { SpotifyAuthService } from "./spotifyauth.service";
 import { ConfigService } from "@nestjs/config";
 
 @Module({
